docker/report-generator/requirements.txt
jinja2==3.1.4
weasyprint==62.3
psycopg2-binary==2.9.9
